Move utils to helper dir
