+++ src/helpers/web3.ts
@@ -1,0 +1,9 @@
+import Web3 from 'web3'
+import { getProvider } from './provider'
+
+import Logger from './Logger'
+
+const log = new Logger('web3')
+log.debug('Init')
+
+export const web3 = new Web3(getProvider())
