+++ src/types/index.ts
@@ -1,1 +1,3 @@
+import BN from 'bn.js'
+
 export type Command = () => void
